// src/api/types/ExtractSchema.ts
/**
 * This file was auto-generated by Fern from our API Definition.
 */

import * as Bytebot from "..";

export type ExtractSchema = Bytebot.ExtractSchema.Text | Bytebot.ExtractSchema.Attribute | Bytebot.ExtractSchema.Table;

export declare namespace ExtractSchema {
    interface Text extends Bytebot.TextExtractSchema {
        type: "Text";
    }

    interface Attribute extends Bytebot.AttributeExtractSchema {
        type: "Attribute";
    }

    interface Table extends Bytebot.TableExtractSchema {
        type: "Table";
    }
}