// src/api/types/ExtractResponseAction.ts
/**
 * This file was auto-generated by Fern from our API Definition.
 */

import * as Bytebot from "..";

export type ExtractResponseAction =
    | Bytebot.ExtractResponseAction.CopyAttribute
    | Bytebot.ExtractResponseAction.CopyText
    | Bytebot.ExtractResponseAction.ExtractTable;

export declare namespace ExtractResponseAction {
    interface CopyAttribute extends Bytebot.CopyAttributeBrowserAction {
        type: "CopyAttribute";
    }

    interface CopyText extends Bytebot.CopyTextBrowserAction {
        type: "CopyText";
    }

    interface ExtractTable extends Bytebot.ExtractTableBrowserAction {
        type: "ExtractTable";
    }
}