// src/api/types/BrowserActionType.ts
/**
 * This file was auto-generated by Fern from our API Definition.
 */

export type BrowserActionType = "CopyAttribute" | "AssignAttribute" | "Click" | "CopyText" | "ExtractTable";

export const BrowserActionType = {
    CopyAttribute: "CopyAttribute",
    AssignAttribute: "AssignAttribute",
    Click: "Click",
    CopyText: "CopyText",
    ExtractTable: "ExtractTable",
} as const;